added feature service and feature object
