+++ service/feature/helper.go
@@ -1,4 +1,4 @@
-package featureset
+package feature
 
 import (
 	"strings"
